requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
numpy==1.21.6
ultralytics==8.0.196
Pillow==9.5.0
aws-wsgi==0.2.7
requests==2.28.1
tqdm==4.64.0
opencv-python-headless==4.5.5.64